refactor: 여행 상태에 따른 분류 계산을 클라이언트 구독 방식으로 변경
- 기존에는 서버 컴포넌트에서 여행 상태(다가오는/지난)를 분류해 계산한 뒤 props로 하위 컴포넌트에 내려줘서 캐시가 무효화돼도 화면에 즉시 반영되지 않았음
- useGetAllUserPlans라는 useSuspenseQuery 기반 훅을 작성해 클라이언트에서 여행 데이터를 직접 구독하도록 수정
- LobbyPlanSection 클라이언트 컴포넌트를 생성하여 여행 데이터를 직접 구독하고, useMemo를 적용한 여행 상태별 분류 계산 로직을 수행
- 로비 페이지(서버 컴포넌트)는 prefetchQuery + HydrationBoundary로 초기 캐싱만 담당하도록 수정
